requirements/base.txt
Django==1.5.1
bpython==0.12
django-braces==1.0.0
django-model-utils==1.3.1
logutils==0.3.3
South==0.7.6
django-toolbelt==0.0.1
PIL==1.1.7
